pyproject.toml
[tool.poetry]
name = "rag-chatbot-sample"
version = "0.1.0"
description = ""
authors = ["k.takeda <52458640+hitohorobe@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <3.13"
streamlit = "^1.36.0"
openai = "^1.35.4"
langchain = "^0.2.6"
langchain-openai = "^0.1.10"
langchain-text-splitters = "^0.2.2"
langchain-pinecone = "^0.1.1"
pinecone-client = "3.2.2"
python-dotenv = "^1.0.1"
langchain-community = "^0.2.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.10"
pytest = "^8.2.2"
mypy = "^1.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
